web-ui/requirements.txt
# Core dependencies
quart==0.18.3
quart-auth
quart-session==1.0.7
werkzeug==2.3.0
jinja2>=3.0.0
markupsafe>=2.0.0
itsdangerous>=2.0.0

# Database and HTTP
asyncpg>=0.27.0
httpx>=0.24.0

# Redis client
redis>=4.5.1
aioredis>=2.0.1

# GraphQL
strawberry-graphql[asgi,debug-server]>=0.171.1
graphql-core>=3.2.3

# Utilities
python-dotenv>=1.0.0
hypercorn>=0.14.4
python-slugify>=8.0.1
shortuuid>=1.0.11
markdown>=3.4.0
pygments>=2.14.0
aiofiles>=23.0.0

# Github integration
pygithub>=2.1.1
gitpython>=3.1.30